Add delete exception

+++ src/main/java/org/first_springboot/first_spriingboot/services/UserService.java
@@ -4,8 +4,10 @@ import org.first_springboot.first_spriingboot.entities.User;
 import org.first_springboot.first_spriingboot.repositories.UserRepository;
 import org.first_springboot.first_spriingboot.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
+import java.util.EmptyStackException;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,7 +30,17 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            userRepository.deleteById(id);
+        } catch (org.springframework.dao.DataIntegrityViolationException e) {
+            throw new org.first_springboot.first_spriingboot.services.exceptions.ResourceNotFoundException(e.getMessage());
+        } catch (RuntimeException e) {
+            e.printStackTrace();
+        }
     }
 
     public User update( Long id, User obj){
